Sample independently from Dirichlet distributions for each of d_k vectors

Sample independently from Dirichlet distributions for each of d_k vectors

Usage

rdirich_dk_vects(nsamp, alpha)

Arguments

nsamp

number of samples to take

alpha

List of vectors of concentration parameters

Value

A matrix with sum(d_k) columns and nsamp rows

Examples

rdirich_dk_vects(10, list(rep(1, 3), rep(1, 4), rep(1, 2)))
#>             [,1]         [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.85903485 0.0007241002 0.14024105 0.16823141 0.10620669 0.67793494
#>  [2,] 0.22777686 0.6923114886 0.07991166 0.22027651 0.74135691 0.03662153
#>  [3,] 0.07009180 0.3913460849 0.53856211 0.30608084 0.04593406 0.54182220
#>  [4,] 0.48592237 0.0816637545 0.43241387 0.01335932 0.32618671 0.35719211
#>  [5,] 0.69731665 0.2395410288 0.06314232 0.16397026 0.04171786 0.11139550
#>  [6,] 0.48061144 0.4502247744 0.06916378 0.29731372 0.08575988 0.48205532
#>  [7,] 0.57273106 0.0511962072 0.37607273 0.22071161 0.05517765 0.35114516
#>  [8,] 0.33541448 0.5772448662 0.08734065 0.46484630 0.11404137 0.39469704
#>  [9,] 0.01942224 0.1338623232 0.84671544 0.21686072 0.24933215 0.08726356
#> [10,] 0.21751343 0.6355332425 0.14695333 0.29840601 0.25367080 0.01826036
#>              [,7]       [,8]       [,9]
#>  [1,] 0.047626966 0.08215247 0.91784753
#>  [2,] 0.001745048 0.87158677 0.12841323
#>  [3,] 0.106162903 0.17893724 0.82106276
#>  [4,] 0.303261861 0.72743711 0.27256289
#>  [5,] 0.682916376 0.51465445 0.48534555
#>  [6,] 0.134871077 0.89833051 0.10166949
#>  [7,] 0.372965583 0.86751847 0.13248153
#>  [8,] 0.026415288 0.79224348 0.20775652
#>  [9,] 0.446543580 0.24827596 0.75172404
#> [10,] 0.429662832 0.95696933 0.04303067